Move UsersController#show -> OauthUsersController
Although this action does somewhat relate to the `User` model, it
doesn't share much in common with the other actions in the
`UsersController` and consequently it was making the code in that
controller more complicated than it needed to be.
